[S07P31D206-204] ã오늘의 미션 생성

+++ BE/greenEarth/src/main/java/com/ssafy/greenEarth/repository/ChildRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ChildRepository extends JpaRepository<Child, Integer> {
 
-    Child findById(Long id);
+    Child findById(int id);
 
 }
