+++ BE/greenEarth/src/main/java/com/ssafy/greenEarth/repository/MissionRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MissionRepository extends JpaRepository<Mission, Integer> {
 
-    Mission findById(Long id);
+    Mission findById(int id);
 
 }
